Fix double deletion issue

+++ lib/fileman/delete.rb
@@ -7,6 +7,7 @@ module Fileman
 	def remove(folder_path)
 		new_path = Fileman.rename(folder_path, 'a', {:include_files => true, :ignore_ext => true, :recursive => true})
 		FileUtils.rm_rf new_path
-		FileUtils.remove_dir new_path
+		# based on different environment, you may need to remove the folder a second time
+		FileUtils.remove_dir new_path if File.exists? new_path
 	end
 end
